% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata2basis.R
\name{fdata2basis2d}
\alias{fdata2basis2d}
\title{base.pc <- create.pc.basis(y1,1:2,norm=F)
bb <- fdata2basis(y1,base.pc,method="inprod")
bb <- fdata2basis(y1,base.pc)
lines(base.pc[[1]])}
\usage{
fdata2basis2d(fdata2d, basis.s, basis.t)
}
\description{
base.pc <- create.pc.basis(y1,1:2,norm=F)
bb <- fdata2basis(y1,base.pc,method="inprod")
bb <- fdata2basis(y1,base.pc)
lines(base.pc[[1]])
}
