% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdata.R
\name{mfdata}
\alias{mfdata}
\alias{c.mfdata}
\alias{is.mfdata}
\alias{names.mfdata}
\alias{[.mfdata}
\alias{plot.mfdata}
\alias{subset.mfdata}
\alias{mfdata.cen}
\alias{mean.mfdata}
\alias{NCOL.mfdata}
\alias{NROW.mfdata}
\alias{ncol.mfdata}
\alias{nrow.mfdata}
\alias{Ops.mfdata}
\alias{Math.mfdata}
\alias{Summary.mfdata}
\title{mfdata class definition and utilities}
\usage{
mfdata(...)

\method{names}{mfdata}(x)

\method{subset}{mfdata}(x, subset, ...)
}
\arguments{
\item{\dots}{Further arguments passed to methods.}

\item{x}{object of class \code{mfdata}}

\item{subset}{subset}
}
\description{
mfdata is a list with fdata type of object:
\itemize{
\item \code{...}  fdata objects of class \code{fdata} with n rows.
}
}
\examples{
data(tecator)
ab0 <- tecator$absorp.fdata
ab1 <- fdata.deriv(ab0)
ab2 <- fdata.deriv(ab0,nderiv=2)
mdat <- mfdata(ab0, ab1, ab2)
is.ldata(mdat)
class(mdat)
plot(mdat[[1]])
plot(mdat[[2]]) 
plot(mdat)
}
