% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights4class.R
\name{weights4class}
\alias{weights4class}
\title{Weighting tools}
\usage{
weights4class(x, type = c("equal", "inverse"))
}
\arguments{
\item{x}{A vector of the labels, true class or observed response. Can be \code{numeric, character, or factor}}

\item{type}{Type of weights.}
}
\description{
computes inverse probability weighting.
}
\seealso{
Other performance: 
\code{\link{accuracy}}
}
\concept{performance}
