% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregre.np.r
\name{fregre.np}
\alias{fregre.np}
\title{Functional regression with scalar response using non-parametric kernel
estimation}
\usage{
fregre.np(
  fdataobj,
  y,
  h = NULL,
  Ker = AKer.norm,
  metric = metric.lp,
  type.S = S.NW,
  par.S = list(w = 1),
  ...
)
}
\arguments{
\item{fdataobj}{\code{\link{fdata}} class object.}

\item{y}{Scalar response with length \code{n}.}

\item{h}{Bandwidth, \code{h>0}. Default argument values are provided as the
5\%--quantile of the distance between \code{fdataobj} curves, see
\code{\link{h.default}}.}

\item{Ker}{Type of asymmetric kernel used, by default asymmetric normal
kernel.}

\item{metric}{Metric function, by default \code{\link{metric.lp}}.}

\item{type.S}{Type of smothing matrix \code{S}. By default \code{S} is
calculated by Nadaraya-Watson kernel estimator (\code{S.NW}).}

\item{par.S}{List of parameters for \code{type.S}: \code{w}, the weights.}

\item{\dots}{Arguments to be passed for \code{\link{metric.lp}} o other
metric function.}
}
\value{
Return:
\itemize{
\item {call}{ The matched call.} 
\item {fitted.values}{ Estimated scalar response.} 
\item {H}{ Hat matrix.} 
\item {residuals}{ \code{y} minus \code{fitted values}.} 
\item {df.residual}{ The residual degrees of freedom.} 
\item {r2}{ Coefficient of determination.} 
\item {sr2}{ Residual variance.} 
\item {y}{ Response.} 
\item {fdataobj}{ Functional explanatory data.} 
\item {mdist}{ Distance matrix between \code{x} and \code{newx}.}
\item {Ker}{ Asymmetric kernel used.} 
\item {h.opt}{ smoothing parameter or' bandwidth.}
}
}
\description{
Computes functional regression between functional explanatory variables and
scalar response using kernel estimation.
}
\details{
The non-parametric functional regression model can be written as follows \deqn{y_i =r(X_i)+\epsilon_i}{ y
= r(X) + \epsilon} where the unknown smooth real function \eqn{r} is
estimated using kernel estimation by means of
\deqn{\hat{r}(X)=\frac{\sum_{i=1}^{n}{K(h^{-1}d(X,X_{i}))y_{i}}}{\sum_{i=1}^{n}{K(h^{-1}d(X,X_{i}))}}}{\hat{r}(X)=(\sum_i
K(d(X,X_i))y_i/h) / (\sum_i K(d(X,X_i)/h)) i=1,...,n} where \eqn{K} is an
kernel function (see \code{Ker} argument), \code{h} is the smoothing
parameter and \eqn{d} is a metric or a semi-metric (see \code{metric}
argument).

The distance between curves is calculated using the \code{\link{metric.lp}}
although any other semimetric could be used (see
\code{\link{semimetric.basis}} or \code{\link{semimetric.NPFDA}} functions).
The kernel is applied to a metric or semi-metrics that provides non-negative
values, so it is common to use asymmetric kernels. Different asymmetric
kernels can be used, see \code{\link{Kernel.asymmetric}}.\cr
}
\examples{
\dontrun{
data(tecator)
absorp=tecator$absorp.fdata
ind=1:129
x=absorp[ind,]
y=tecator$y$Fat[ind]

res.np=fregre.np(x,y,Ker=AKer.epa)
summary(res.np)
res.np2=fregre.np(x,y,Ker=AKer.tri)
summary(res.np2)

# with other semimetrics.
res.pca1=fregre.np(x,y,Ker=AKer.tri,metri=semimetric.pca,q=1)
summary(res.pca1)
res.deriv=fregre.np(x,y,metri=semimetric.deriv)
summary(res.deriv)
x.d2=fdata.deriv(x,nderiv=1,method="fmm",class.out='fdata')
res.deriv2=fregre.np(x.d2,y)
summary(res.deriv2)
x.d3=fdata.deriv(x,nderiv=1,method="bspline",class.out='fdata')
res.deriv3=fregre.np(x.d3,y)
summary(res.deriv3)
}

}
\references{
Ferraty, F. and Vieu, P. (2006). \emph{Nonparametric functional
data analysis.} Springer Series in Statistics, New York. \cr

Febrero-Bande, M., Oviedo de la Fuente, M. (2012).  \emph{Statistical
Computing in Functional Data Analysis: The R Package fda.usc.} Journal of
Statistical Software, 51(4), 1-28. \url{http://www.jstatsoft.org/v51/i04/}

Hardle, W. \emph{Applied Nonparametric Regression}. Cambridge University
Press, 1994.
}
\seealso{
See Also as: \code{\link{fregre.np.cv}},
\code{\link{summary.fregre.fd}} and \code{\link{predict.fregre.fd}} .\cr
Alternative method: \code{\link{fregre.basis}},cand \code{\link{fregre.pc}}.
}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{regression}
