% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phoneme}
\alias{phoneme}
\title{phoneme data}
\format{Elements of phoneme:\cr \code{..$learn}: learning sample of curves.
\code{fdata} class object with: i.- \code{"data"}: Matrix of class
\code{fdata} with 250 curves (rows) discretized in 150 points or argvals
(columns).\cr, ii.- \code{"argvals"}, iii.- \code{"rangeval"}:
range(\code{"argvals"}), iv.- \code{"names"} list with: \code{main} an
overall title "Phoneme learn", \code{xlab} title for \code{x} axis
"frequencies" and \code{ylab} title for \code{y} axis "log-periodograms".\cr
\cr \code{..$test}: testing sample of curves. \code{fdata} class object
with: i.- \code{"data"}: Matrix of class \code{fdata} with 250 curves (rows)
discretized in 150 points or argvals (columns).\cr, ii.- \code{"argvals"},
iii.- \code{"rangeval"}: range(\code{"argvals"}), iv.- \code{"names"} list
with: \code{main} an overall title "Phoneme learn", \code{xlab} title for
\code{x} axis "frequencies" and \code{ylab} title for \code{y} axis
"log-periodograms".\cr \cr \code{..$classlearn}:learning class numbers (as
factor). Factor levels: "sh" 1, "iy" 2, "dcl" 3, "aa" 4 and "ao" 5.\cr \cr
\code{..$classtest}: testing class numbers (as factor). Factor levels: "sh"
1, "iy" 2, "dcl" 3, "aa" 4 and "ao" 5.\cr}
\source{
\url{http://www.math.univ-toulouse.fr/staph/npfda/npfda-datasets.html}
}
\description{
Phoneme curves
}
\details{
The following instructions have been used file: \cr
\url{http://www.math.univ-toulouse.fr/staph/npfda/npfda-phondiscRS.txt}\cr
of \code{Phoneme dataset} file.
}
\examples{

data(phoneme)
names(phoneme)
names(phoneme$learn)
class(phoneme$learn)
dim(phoneme$learn)
table(phoneme$classlearn)

}
\references{
Ferraty, F. and Vieu, P. (2006). \emph{NPFDA in practice}. Free
access on line at \url{http://www.lsp.ups-tlse.fr/staph/npfda/}
}
\author{
Manuel Febrero-Bande and Manuel Oviedo de la Fuente
<manuel.oviedo@usc.es>
}
\keyword{datasets}
