% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rproc2fdata.R
\name{rproc2fdata}
\alias{rproc2fdata}
\title{Simulate several random processes.}
\usage{
rproc2fdata(
  n,
  t = NULL,
  mu = rep(0, length(t)),
  sigma = 1,
  par.list = list(scale = 1, theta = 0.2 * diff(rtt), H = 0.5),
  norm = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n}{Number of functional curves to be generated.}

\item{t}{Discretization points.}

\item{mu}{\code{vector} which specifies the trend values at the
discretization points, by default \code{mu}=\eqn{\mu(t)=0}.  If \code{mu} is
a \code{fdata} class object, \code{t}\eqn{=}\code{argvals(mu)}.}

\item{sigma}{A positive-definite symmetric matrix,
\eqn{\Sigma_{s,t}}{\Sigma}, specifying the covariance matrix among grid
points.  If \code{sigma} is a \code{scalar}, creates a random Gaussian
process with \eqn{\Sigma_{s,t}=}{\Sigma=}\code{sigmaI} (by default
\code{sigma=1}).\cr If \code{sigma} is a \code{vector}, creates a random
Gaussian process with \eqn{\Sigma_{s,t}=}{\Sigma=}\code{diag(sigma)}.\cr If
\code{sigma} is a character: create a random process using the covariance
matrix \eqn{\Sigma_{s,t}}{\Sigma} indicated in the argument, \itemize{ \item
\code{"OU"} or \code{"OrnsteinUhlenbeck"}, creates a random Ornstein
Uhlenbeck process with
\eqn{\Sigma_{s,t}=\frac{\sigma^2}{2\theta}e^{-\theta\left(s+t\right)}
\left(e^{2\theta\left(s+t\right)}-1\right)}{\Sigma(s,t)=\sigma^2/\theta
exp(-\theta(s+t))(exp(2\theta(s+t)-1))}, by default
\eqn{\theta=1/(3range(t))}{\theta=1/(3range(t))},
\eqn{\sigma^2={1}}{\sigma^2=1}.  \item \code{"brownian"} or \code{"wiener"},
creates a random Wiener process with \eqn{\Sigma_{s,t}=\sigma^2
min(s,t)}{\Sigma(s,t)=\sigma^2 min(s,t)}, by default
\eqn{\sigma^2=1}{\sigma^2=1}. \item \code{"fbrownian"}, creates a random
fractional brownian process with
\eqn{\Sigma_{s,t}=\sigma^{2H}/2{|s|^{2H}+|t|^{2H}-|s-t|^{2H}}}{\Sigma(s,t)=\sigma^{2H}/2{|s|^{2H}+|t|^{2H}-|s-t|^{2H}}},
by default \eqn{\sigma^2=1}{\sigma^2=1} and \eqn{H=0.5}{H=0.5} (brownian
process).  \item \code{"vexponential"}, creates a random gaussian process
with exponential variogram \eqn{\Sigma_{s,t}=\sigma^2
e^{\left(-\frac{\left|s-t\right|}{\theta}\right)}}{\Sigma=\sigma^2
exp(-|s-t|/\theta )}, by default \eqn{\theta={0.2
range(t)}}{\theta=.2*range(t)}, \eqn{\sigma^2={1}}{\sigma^2=1}. }}

\item{par.list}{List of parameter to process, by default \code{"scale"}
\eqn{\sigma^2=1}{\sigma^2=1}, \code{"theta"} \eqn{\theta=0.2
range(t)}{\theta=0.2 range(t)} and \code{"H"}=0.5.}

\item{norm}{If \code{TRUE} the norm of random projection is 1. Default is
\code{FALSE}}

\item{verbose}{If \code{TRUE}, information about procedure is printed.
Default is \code{FALSE}.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Return the functional random processes as a \code{fdata} class
object.
}
\description{
Simulate Functional Data from different processes: Ornstein Uhlenbeck,
Brownian, Fractional Brownian, Gaussian or Exponential variogram.
}
\examples{
\dontrun{
par(mfrow=c(3,2))
lent<-30
tt<-seq(0,1,len=lent)
mu<-fdata(rep(0,lent),tt)
plot(rproc2fdata(200,t=tt,sigma="OU",par.list=list("scale"=1)))
plot(rproc2fdata(200,mu=mu,sigma="OU",par.list=list("scale"=1)))
plot(rproc2fdata(200,t=tt,sigma="vexponential"))
plot(rproc2fdata(200,t=tt,sigma=1:lent))
plot(rproc2fdata(200,t=tt,sigma="brownian"))
plot(rproc2fdata(200,t=tt,sigma="wiener"))
#plot(rproc2fdata(200,seq(0,1,len=30),sigma="oo")) # this is an error 
}

}
\author{
Manuel Febrero-Bande, Manuel Oviedo de la Fuente
\email{manuel.oviedo@udc.es}
}
\keyword{datagen}
