% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.lfdata.R
\name{subset.fdata}
\alias{subset.fdata}
\title{Subsetting}
\usage{
\method{subset}{fdata}(x, subset, select, drop = TRUE, ...)
}
\arguments{
\item{x}{object to be subsetted (\code{fdata} class).}

\item{subset}{logical expression indicating elements or rows to keep.}

\item{select}{logical expression indicating points or columns to keep.}

\item{drop}{passed on to \code{[} indexing operator.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
An object similar to \code{x} contain just the selected elements.
}
\description{
Return subsets of \code{fdata} which meet conditions.
}
\seealso{
See \code{\link{subset}} and \code{\link{fdata}}.
}
