% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tecator}
\alias{tecator}
\title{tecator data}
\format{
The format is: \cr \code{..$absorp.fdata}: absorbance data.
\code{fdata} class object with: \cr \itemize{ \item \code{"data"}: Matrix of
class \code{fdata} with 215 curves (rows) discretized in 100 points or
argvals (columns).\cr \item \code{"argvals"}: 100 discretization points from
850 to 1050nm \cr \item \code{"rangeval"}=(850,1050):
range(\code{"argvals"}) \item \code{"names"} list with: \code{main} an
overall title "Tecator data set", \code{xlab} title for \code{x} axis
"Wavelength (nm)" and \code{ylab} title for \code{y} axis "Absorbances". }
\code{..$y}: the percentages of Fat, Water and Protein.  The three contents
are determined by analytic chemistry.\cr
}
\description{
Water, Fat and Protein content of meat samples
}
\details{
\code{absorp.fdata} absorbance data for 215 samples. The first 129 were
originally used as a training set endpoints the percentages of Fat, Water
and Protein.\cr for more details see tecator package
}
\examples{
data(tecator)
names(tecator)
names(tecator$absorp.fdata)
names(tecator$y)
names(tecator$y)
class(tecator$absorp.fdata)
class(tecator$y)
dim(tecator$absorp.fdata)
dim(tecator$y)

}
\author{
Manuel Febrero-Bande and Manuel Oviedo de la Fuente \email{manuel.oviedo@udc.es}
}
\keyword{datasets}
